[MODEL-8479] DRUM REFACTOR - Move utils into separate files
