+++ custom_model_runner/datarobot_drum/drum/utils/drum_utils.py
@@ -1,34 +1,14 @@
-"""
-Copyright 2021 DataRobot, Inc. and its affiliates.
-All rights reserved.
-This is proprietary source code of DataRobot, Inc. and its affiliates.
-Released under the terms of DataRobot Tool and Utility Agreement.
-"""
-import io
 import json
-import logging
 import os
 import socket
-from contextlib import closing, contextmanager
+from contextlib import closing
 from functools import partial
 from pathlib import Path
 
-import numpy as np
-import pandas as pd
-from jinja2 import BaseLoader, DebugUndefined, Environment
-from scipy.io import mmread
+from jinja2 import Environment, BaseLoader, DebugUndefined
 
-from datarobot_drum.drum.common import get_pyarrow_module
-from datarobot_drum.drum.enum import (
-    ArgumentOptionsEnvVars,
-    InputFormatExtension,
-    InputFormatToMimetype,
-    LOGGER_NAME_PREFIX,
-    PredictionServerMimetypes,
-)
-from datarobot_drum.drum.exceptions import DrumCommonException
-
-logger = logging.getLogger(LOGGER_NAME_PREFIX + "." + __name__)
+from datarobot_drum.drum.enum import ArgumentOptionsEnvVars
+from datarobot_drum.drum.utils import logger
 
 
 class DrumUtils:
@@ -190,94 +170,3 @@ def unset_drum_supported_env_vars(additional_unset_vars=[]):
         ArgumentOptionsEnvVars.VALUE_VARS + ArgumentOptionsEnvVars.BOOL_VARS + additional_unset_vars
     ):
         os.environ.pop(env_var_key, None)
-
-
-class StructuredInputReadUtils:
-    @staticmethod
-    def read_structured_input_file_as_binary(filename):
-        mimetype = StructuredInputReadUtils.resolve_mimetype_by_filename(filename)
-        with open(filename, "rb") as file:
-            binary_data = file.read()
-        return binary_data, mimetype
-
-    @staticmethod
-    def read_structured_input_file_as_df(filename, sparse_column_file=None):
-        binary_data, mimetype = StructuredInputReadUtils.read_structured_input_file_as_binary(
-            filename
-        )
-        if sparse_column_file:
-            with open(sparse_column_file, "rb") as file:
-                sparse_colnames = file.read()
-        else:
-            sparse_colnames = None
-        return StructuredInputReadUtils.read_structured_input_data_as_df(
-            binary_data, mimetype, sparse_colnames
-        )
-
-    @staticmethod
-    def resolve_mimetype_by_filename(filename):
-        return InputFormatToMimetype.get(os.path.splitext(filename)[1])
-
-    @staticmethod
-    def read_structured_input_data_as_df(binary_data, mimetype, sparse_colnames=None):
-        try:
-            if mimetype == PredictionServerMimetypes.TEXT_MTX:
-                columns = None
-                if sparse_colnames:
-                    columns = [
-                        column.strip().decode("utf-8")
-                        for column in io.BytesIO(sparse_colnames).readlines()
-                    ]
-                return pd.DataFrame.sparse.from_spmatrix(
-                    mmread(io.BytesIO(binary_data)), columns=columns
-                )
-            elif mimetype == PredictionServerMimetypes.APPLICATION_X_APACHE_ARROW_STREAM:
-                df = get_pyarrow_module().ipc.deserialize_pandas(binary_data)
-
-                # After CSV serialization+deserialization,
-                # original dataframe's None and np.nan values
-                # become np.nan values.
-                # After Arrow serialization+deserialization,
-                # original dataframe's None and np.nan values
-                # become np.nan for numeric columns and None for 'object' columns.
-                #
-                # Since we are supporting both CSV and Arrow,
-                # to be consistent with CSV serialization/deserialization,
-                # it is required to replace all None with np.nan for Arrow.
-                df.fillna(value=np.nan, inplace=True)
-
-                return df
-            else:  # CSV format
-                df = pd.read_csv(io.BytesIO(binary_data))
-
-                # If the DataFrame only contains a single column, treat blank lines as NANs
-                if df.shape[1] == 1:
-                    logger.info(
-                        "Input data only contains a single column, treating blank lines as NaNs"
-                    )
-                    df = pd.read_csv(io.BytesIO(binary_data), skip_blank_lines=False)
-
-                return df
-
-        except pd.errors.ParserError as e:
-            raise DrumCommonException(
-                "Pandas failed to read input binary data {}".format(binary_data)
-            )
-
-
-@contextmanager
-def capture_R_traceback_if_errors(r_handler, logger):
-    from rpy2.rinterface_lib.embedded import RRuntimeError
-
-    try:
-        yield
-    except RRuntimeError as e:
-        try:
-            out = "\n".join(r_handler("capture.output(traceback(max.lines = 50))"))
-            logger.error("R Traceback:\n{}".format(str(out)))
-        except Exception as traceback_exc:
-            e.context = {
-                "r_traceback": "(an error occurred while getting traceback from R)",
-                "t_traceback_err": traceback_exc,
-            }
-        raise
