+++ tests/drum/test_deployment_config.py
@@ -31,7 +31,7 @@ from tests.drum.constants import (
 from datarobot_drum.resource.utils import _create_custom_model_dir
 
 from datarobot_drum.resource.drum_server_utils import DrumServerRun
-from datarobot_drum.drum.utils import unset_drum_supported_env_vars
+from datarobot_drum.drum.utils.drum_utils import unset_drum_supported_env_vars
 
 
 class TestDeploymentConfig:
