+++ tests/drum/test_stats.py
@@ -22,7 +22,7 @@ from .constants import (
 from datarobot_drum.resource.drum_server_utils import DrumServerRun
 from datarobot_drum.resource.utils import _create_custom_model_dir
 
-from datarobot_drum.drum.utils import unset_drum_supported_env_vars
+from datarobot_drum.drum.utils.drum_utils import unset_drum_supported_env_vars
 
 
 class TestInference:
